Enum rcue::cue::Command

pub enum Command {
    Rem(String, String),
    Catalog(String),
    CdTextFile(String),
    Flags(Vec<String>),
    Isrc(String),
    Songwriter(String),
    Performer(String),
    Title(String),
    File(String, String),
    Track(String, String),
    Index(String, String),
    Pregap(String),
    Postgap(String),
    Unknown(String),
    None,
}

Represents a CUE command in a CUE sheet.

Variants
Rem(String, String)

ignore comment

Catalog(String)

Media Catalog Number (MCN) of the disc

CdTextFile(String)

Path to the file containing the CD-Text meta-data of the disc

Flags(Vec<String>)

Special sub-code flags for a track

Isrc(String)

The International Standard Recording Code for a track, typically CCOOOYYSSSSS

C: Country code (uppercase alphanumeric) O: Owner code (uppercase alphanumeric) Y: Year (numeric) S: Serial number (numeric)

Songwriter(String)

Songwriter for a disc or a track

Performer(String)

Performer for a disc or a track

Title(String)

Title for a disc or a track

File(String, String)

(file name, file type) of a disc, to be used by tracks

Track(String, String)

A track on a disc

Index(String, String)

Defines the index of a track on the disc

Pregap(String)

Length of a track's pregap, (mm:ss:ff) where each frame is 1/75 of a second

Postgap(String)

Length of a track's postgap (mm:ss:ff) where each frame is 1/75 of a second

Unknown(String)

Unknown command

None

Not a command