Function rcue::parser::parse

pub fn parse(buf_reader: &mut BufRead, strict: bool) -> Result<Cue, CueError>

Parses a BufRead into a Cue struct.

Strict mode will return CueError if invalid fields or extra lines are detected. When not in strict mode, bad lines and fields will be skipped, and unknown fields will be stored in Cue.unknown.

Example

use rcue::parser::parse;
use std::fs::File;
use std::io::BufReader;

let file = File::open("test/fixtures/unicode.cue").unwrap();
let mut buf_reader = BufReader::new(file);
let cue = parse(&mut buf_reader, true).unwrap();
assert_eq!(cue.title, Some("マジコカタストロフィ".to_string()));

Failures

Fails if the CUE file can not be parsed.